Moved authController into auth
